// apps/web/src/components/pages/landing/JaanPanahPage/enum.ts
export enum JaanPanahGenderEnum {
  None = 0,
  Male = 1,
  Female = 2,
}

export enum JaanPanahMarriageEnum {
  None = 0,
  Single = 1,
  Married = 2,
}

export enum JaanPanahHaveChildrenEnum {
  None = 0,
  Yes = 1,
  No = 2,
}

export enum JaanPanahCityConditionEnum {
  None = 0,
  Danger = 1,
  Normal = 2,
}